requirements.txt
apache-airflow==2.7.3
apache-airflow-providers-snowflake==4.4.0
apache-airflow-providers-google==10.13.1
snowflake-sqlalchemy==1.5.1
snowflake-connector-python==3.8.0
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
pyarrow==14.0.1
dbt-core==1.7.9
dbt-snowflake==1.7.5